% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_test.R
\name{knn_test}
\alias{knn_test}
\title{KNN test}
\usage{
knn_test(geo, time, B, K)
}
\arguments{
\item{geo}{A dataframe with two columns containing geographic(coordinates) data.}

\item{time}{A vector containing time data}

\item{B}{Number of permutations for Monte Carlo permutation test}

\item{K}{the number of k nearest neightbors}
}
\value{
A B*2 dataframe composed of two test statistics D_k, lambda_D_k.
}
\description{
Spatiotemporal k Nearest Neighbors Test with Monte Carlo permutation test
}
\examples{
geo <- matrix(rnorm(1000 * 2), 1000, 2)
time <- rexp(1000)
res = knntest(geo,time,B=999,K=5)
}
