% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knoxtest.R
\name{knoxtest}
\alias{knoxtest}
\title{Knox test}
\usage{
knoxtest(geo, time, delta = NULL, gamma = NULL, delta_ratio, gamma_ratio, B)
}
\arguments{
\item{geo}{A dataframe with two columns containing geographic(coordinates) data.}

\item{time}{A vector containing time data}

\item{delta}{A numeric value of critical space distances values}

\item{gamma}{A numeric value of critical time distances values}

\item{delta_ratio}{Optional, select delta as (1-a)th percentile of space distances where a is delta_ratio}

\item{gamma_ratio}{Optional, select delta as (1-b)th percentile of time distances where b is gamma_ratio}

\item{B}{Number of permutations for Monte Carlo permutation test}
}
\value{
A list of randomized knox test statistic, observed knox text statistic, p value from Monte Carlo permutation test
}
\description{
Knox test with Monte Carlo permutation test
}
\examples{
geo <- matrix(rnorm(1000 * 2), 1000, 2)
time <- rexp(1000)
res1 = knoxtest(geo,time,delta=50,gamma=30,B=999)
res1 = knoxtest(geo,time,delta_ratio=0.1,gamma_ratio=0.02,B=999)
}
