% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manteltest.R
\name{manteltest}
\alias{manteltest}
\title{Mantel test}
\usage{
manteltest(geo, time, B)
}
\arguments{
\item{geo}{A dataframe with two columns containing geographic(coordinates) data.}

\item{time}{A vector containing time data}

\item{B}{Number of permutations for Monte Carlo permutation test}
}
\value{
A list of randomized Mantel test statistic, observed Mantel text statistic, p value from Monte Carlo permutation test
}
\description{
Mantel test with Monte Carlo permutation test
}
\examples{
library(surveillance)
data("imdepi")
imdepiB <- subset(imdepi, type == "B")
g = coordinates(imdepiB$events)
t = imdepiB$events$time
res1 = manteltest(g,t,999)
}
