% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stcls_hyp_plot.R
\name{stcls_hyp_plot}
\alias{stcls_hyp_plot}
\title{Comparison between knox, mantel, knn test}
\usage{
stcls_hyp_plot(
  geo,
  time,
  delta = NULL,
  gamma = NULL,
  delta_ratio,
  gamma_ratio,
  B,
  K
)
}
\arguments{
\item{geo}{A dataframe with two columns containing geographic(coordinates) data.}

\item{time}{A vector containing time data}

\item{delta}{A numeric value of critical space distances values}

\item{gamma}{A numeric value of critical time distances values}

\item{delta_ratio}{Optional, select delta as (1-a)th percentile of space distances where a is delta_ratio}

\item{gamma_ratio}{Optional, select delta as (1-b)th percentile of time distances where b is gamma_ratio}

\item{B}{Number of permutations for Monte Carlo permutation test}

\item{K}{the number of k nearest neightbors}
}
\value{
4 histograms of know test, mantel test, knn test with cumulative test statistic, knn test with k-specific test statistic with p-value of Monte Carlo permutation test respectively
}
\description{
Compare the hypothesis testing-based clustering methods of the events spatio-temporal data type by histograms with p-value of Monte Carlo permutation test
}
\examples{
geo <- matrix(rnorm(1000 * 2), 1000, 2)
time <- rexp(1000)
stcls_hyp_plot(geo, time, delta=50, gamma=30, B=10,K=5)
stcls_hyp_plot(geo, time, delta_ratio = 0.1, gamma_ratio = 0.02, B=10,K=5)
}
