% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_test.R
\name{knn_test}
\alias{knn_test}
\title{KNN test}
\usage{
knn_test(geo, time, B, K)
}
\arguments{
\item{geo}{A dataframe with two columns containing geographic(coordinates) data.}

\item{time}{A vector containing time data}

\item{B}{Number of permutations for Monte Carlo permutation test}

\item{K}{the number of k nearest neightbors}
}
\value{
A B*2 dataframe composed of two test statistics D_k, lambda_D_k.
}
\description{
Spatiotemporal k Nearest Neighbors Test with Monte Carlo permutation test
}
\examples{
library(surveillance)
data("imdepi")
imdepiB <- subset(imdepi, type == "B")
g = coordinates(imdepiB$events)
t = imdepiB$events$time
res = knntest(g,t,B=999,K=5)
}
