% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_test_stat.R
\name{knn_test_stat}
\alias{knn_test_stat}
\title{KNN test statistics}
\usage{
knn_test_stat(geo, time, K)
}
\arguments{
\item{geo}{A dataframe with two columns containing geographic(coordinates) data.}

\item{time}{A vector containing time data}

\item{K}{the number of k nearest neightbors}
}
\value{
A vector of test statistics c(D_k, lambda_D_k)
}
\description{
Calculate cumulative test statistic and k– specific test statistic of spatiotemporal k Nearest Neighbors Test.
}
\examples{
library(surveillance)
data("imdepi")
imdepiB <- subset(imdepi, type == "B")
g = coordinates(imdepiB$events)
t = imdepiB$events$time
res = knn_test_stat(g,t,K=5)
}
