% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manteltest.R
\name{manteltest}
\alias{manteltest}
\title{Mantel test}
\usage{
manteltest(geo, time, B)
}
\arguments{
\item{geo}{A dataframe with two columns containing geographic(coordinates) data.}

\item{time}{A vector containing time data}

\item{B}{Number of permutations for Monte Carlo permutation test}
}
\value{
A list of randomized Mantel test statistic, observed Mantel text statistic, p value from Monte Carlo permutation test
}
\description{
Mantel test with Monte Carlo permutation test
}
\examples{
geo <- matrix(rnorm(1000 * 2), 1000, 2)
time <- rexp(1000)
res1 = manteltest(geo,time,999)
}
